[ANDROSDK-779] Extend APIErrorMapper
